requirements.txt
flask
flask_sqlalchemy
psycopg2
psycopg2-binary